+++ packages/components/accordion/src/accordion.tsx
@@ -1,4 +1,5 @@
 import {forwardRef} from "@nextui-org/system";
+import {LayoutGroup} from "framer-motion";
 
 import {UseAccordionProps, useAccordion} from "./use-accordion";
 import {AccordionProvider} from "./accordion-context";
@@ -7,23 +8,33 @@ import AccordionItem from "./accordion-item";
 export interface AccordionProps extends Omit<UseAccordionProps, "ref"> {}
 
 const AccordionGroup = forwardRef<AccordionProps, "div">((props, ref) => {
-  const {Component, context, state, getBaseProps, handleFocusChanged, itemStyles} = useAccordion({
+  const {
+    Component,
+    context,
+    state,
+    getBaseProps,
+    disableAnimation,
+    handleFocusChanged,
+    itemStyles,
+  } = useAccordion({
     ref,
     ...props,
   });
 
+  const content = [...state.collection].map((item) => (
+    <AccordionItem
+      key={item.key}
+      item={item}
+      onFocusChange={(isFocused) => handleFocusChanged(isFocused, item.key)}
+      {...item.props}
+      classNames={{...itemStyles, ...(item.props.classNames || {})}}
+    />
+  ));
+
   return (
     <Component {...getBaseProps()}>
       <AccordionProvider value={context}>
-        {[...state.collection].map((item) => (
-          <AccordionItem
-            key={item.key}
-            item={item}
-            onFocusChange={(isFocused) => handleFocusChanged(isFocused, item.key)}
-            {...item.props}
-            classNames={{...itemStyles, ...(item.props.classNames || {})}}
-          />
-        ))}
+        {disableAnimation ? content : <LayoutGroup>{content}</LayoutGroup>}
       </AccordionProvider>
     </Component>
   );