+++ packages/components/badge/src/use-badge.ts
@@ -44,7 +44,11 @@ export function useBadge(originalProps: UseBadgeProps) {
 
   const domRef = useDOMRef(ref);
 
-  const isOneChar = useMemo(() => String(content)?.length === 1, [content]);
+  const isOneChar = useMemo(() => String(content)?.length === 1 || originalProps?.isOneChar, [
+    content,
+    originalProps?.isOneChar,
+  ]);
+
   const isDot = useMemo(() => String(content)?.length === 0, [content]);
 
   const baseStyles = clsx(classNames?.badge, className);
@@ -74,6 +78,8 @@ export function useBadge(originalProps: UseBadgeProps) {
     content,
     slots,
     classNames,
+    disableAnimation: originalProps?.disableAnimation,
+    isInvisible: originalProps?.isInvisible,
     getBadgeProps,
   };
 }