+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @nextui-org/popover
 
+## 0.0.0-dev-v2-20230521141624
+
+### Patch Changes
+
+- Badge improved
+- Updated dependencies
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230521141624
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230521141624
+  - @nextui-org/use-aria-button@0.0.0-dev-v2-20230521141624
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230521141624
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230521141624
+  - @nextui-org/button@0.0.0-dev-v2-20230521141624
+  - @nextui-org/system@0.0.0-dev-v2-20230521141624
+  - @nextui-org/theme@0.0.0-dev-v2-20230521141624
+
 ## 0.0.0-dev-v2-20230518012432
 
 ### Patch Changes